feat(auth-3): 기본 회원가입 및 로그인 여정 개발
- 이름이 바뀐 프로덕트들을 푸시하는 커밋
