--- src/main/java/com/stamp/global/util/JwtProvider.java
@@ -1,44 +1,0 @@
-package com.stamp.global.util;
-
-import com.stamp.api.auth.domain.entity.EmployerUser;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-import javax.crypto.SecretKey;
-import org.springframework.stereotype.Component;
-
-@Component
-public class JwtProvider {
-    private final SecretKey secretKey;
-
-    public JwtProvider() {
-        String secret = "your-secret-key-that-is-at-least-32-bytes-long"; // 32바이트 이상 필요
-        byte[] keyBytes = secret.getBytes(StandardCharsets.UTF_8);
-        this.secretKey = Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    public String generateToken(EmployerUser member) {
-        long expirationMs = 1000 * 60 * 120;
-        return Jwts.builder()
-                .subject(String.valueOf(member.getId()))
-                .claim("email", member.getEmail())
-                .claim("contact", member.getContact())
-                .claim("roles", "Employer")
-                .issuedAt(new Date())
-                .expiration(new Date(System.currentTimeMillis() + expirationMs))
-                .signWith(secretKey)
-                .compact();
-    }
-
-    public String validateAndGetUserId(String token) {
-        Claims claims = Jwts.parser()
-                .verifyWith(secretKey)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload();
-
-        return claims.getSubject();
-    }
-}
